refactor: move disable operator functionality
