+++ packages/apps/human-app/server/src/integrations/reputation-oracle/reputation-oracle.interface.ts
@@ -4,7 +4,7 @@ import { SigninWorkerData } from '../../modules/user-worker/model/worker-signin.
 import { EmailVerificationData } from '../../modules/email-confirmation/model/email-verification.model';
 import { ResendEmailVerificationData } from '../../modules/email-confirmation/model/resend-email-verification.model';
 import { PrepareSignatureData } from '../../modules/prepare-signature/model/prepare-signature.model';
-import { DisableOperatorData } from '../../modules/disable-operator/model/disable-operator.model';
+import { DisableOperatorData } from '../../modules/user-operator/model/disable-operator.model';
 import { RegisterAddressData } from '../../modules/register-address/model/register-address.model';
 import { RestorePasswordData } from '../../modules/password-reset/model/restore-password.model';
 import { TokenRefreshData } from '../../modules/token-refresh/model/token-refresh.model';
