--- packages/apps/human-app/server/src/modules/disable-operator/spec/disable-operator.service.mock.ts
@@ -1,3 +1,0 @@
-export const serviceMock = {
-  processDisableOperator: jest.fn(),
-};
