+++ packages/apps/human-app/server/src/modules/user-operator/operator.controller.ts
@@ -1,3 +1,5 @@
+import { Mapper } from '@automapper/core';
+import { InjectMapper } from '@automapper/nestjs';
 import {
   Body,
   Controller,
@@ -5,10 +7,11 @@ import {
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
+import { ApiBearerAuth, ApiOperation, ApiTags } from '@nestjs/swagger';
+
+import { Authorization } from '../../common/config/params-decorators';
+
 import { OperatorService } from './operator.service';
-import { Mapper } from '@automapper/core';
-import { InjectMapper } from '@automapper/nestjs';
-import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import {
   SignupOperatorCommand,
   SignupOperatorDto,
@@ -18,6 +21,10 @@ import {
   SigninOperatorDto,
   SigninOperatorResponse,
 } from './model/operator-signin.model';
+import {
+  DisableOperatorCommand,
+  DisableOperatorDto,
+} from './model/disable-operator.model';
 
 @Controller()
 export class OperatorController {
@@ -29,7 +36,7 @@ export class OperatorController {
   @Post('/auth/web3/signup')
   @ApiOperation({ summary: 'Operator signup' })
   @UsePipes(new ValidationPipe())
-  public signupOperator(
+  async signupOperator(
     @Body() signupOperatorDto: SignupOperatorDto,
   ): Promise<void> {
     const signupOperatorCommand = this.mapper.map(
@@ -44,7 +51,7 @@ export class OperatorController {
   @Post('/auth/web3/signin')
   @ApiOperation({ summary: 'Operator signin' })
   @UsePipes(new ValidationPipe())
-  public signinOperator(
+  async signinOperator(
     @Body() dto: SigninOperatorDto,
   ): Promise<SigninOperatorResponse> {
     const command = this.mapper.map(
@@ -53,5 +60,25 @@ export class OperatorController {
       SigninOperatorCommand,
     );
     return this.service.signinOperator(command);
+  }
+
+  @ApiTags('User-Operator')
+  @Post('/disable-operator')
+  @ApiOperation({
+    summary: 'Endpoint to disable an operator',
+  })
+  @ApiBearerAuth()
+  @UsePipes(new ValidationPipe())
+  async disableOperator(
+    @Body() disableOperatorDto: DisableOperatorDto,
+    @Authorization() token: string,
+  ): Promise<void> {
+    const disableOperatorCommand = this.mapper.map(
+      disableOperatorDto,
+      DisableOperatorDto,
+      DisableOperatorCommand,
+    );
+    disableOperatorCommand.token = token;
+    return this.service.processDisableOperator(disableOperatorCommand);
   }
 }
