+++ packages/apps/human-app/server/src/modules/user-operator/operator.service.ts
@@ -1,7 +1,10 @@
 import { Injectable } from '@nestjs/common';
-import { SignupOperatorCommand } from './model/operator-registration.model';
+
 import { ReputationOracleGateway } from '../../integrations/reputation-oracle/reputation-oracle.gateway';
+
 import { SigninOperatorCommand } from './model/operator-signin.model';
+import { SignupOperatorCommand } from './model/operator-registration.model';
+import { DisableOperatorCommand } from './model/disable-operator.model';
 
 @Injectable()
 export class OperatorService {
@@ -13,5 +16,9 @@ export class OperatorService {
 
   signinOperator(command: SigninOperatorCommand) {
     return this.gateway.sendOperatorSignin(command);
+  }
+
+  processDisableOperator(command: DisableOperatorCommand): Promise<void> {
+    return this.gateway.sendDisableOperator(command);
   }
 }
